refactored theme
