+++ packages/ui-components/src/theme/config.ts
@@ -1,0 +1,8 @@
+import { ThemeConfig } from '@chakra-ui/react';
+
+const config: ThemeConfig = {
+  initialColorMode: 'light',
+  useSystemColorMode: false,
+};
+
+export default config;
