+++ packages/ui-components/src/theme/fontSizes.ts
@@ -1,0 +1,6 @@
+const fontSizes = {
+  '2xs': '10px',
+  '3xs': '8px',
+};
+
+export default fontSizes;
