+++ packages/ui-components/src/theme/fonts.ts
@@ -1,0 +1,6 @@
+const fonts = {
+  heading: "'Inter', sans-serif",
+  body: "'Inter', sans-serif",
+};
+
+export default fonts;
