+++ packages/ui-components/src/theme/styles.ts
@@ -1,0 +1,17 @@
+import { Styles } from '@chakra-ui/theme-tools';
+
+const styles: Styles = {
+  global: {
+    html: {
+      scrollBehavior: 'smooth',
+    },
+    body: {
+      minH: '100vh',
+      color: 'gray.800',
+      bgColor: 'white',
+      scrollBehavior: 'smooth',
+    },
+  },
+};
+
+export default styles;
