+++ packages/ui-components/src/theme/zIndices.ts
@@ -1,0 +1,6 @@
+const zIndices = {
+  commentDrawer: 1450,
+  commentOverlay: 1400,
+};
+
+export default zIndices;
