--- packages/ui-components/src/theme/zindex.ts
@@ -1,4 +1,0 @@
-export const zIndices = {
-  commentDrawer: 1450,
-  commentOverlay: 1400,
-}
